Global optimizations: 	- Optimization of pixel storage to avoid multiple functions 	- Optimization of fx calculations, to minimize them 	- Optimization of background, initialized one single time, then copied 	- Optimization of storage of the line of pixels 	- Optimization of texturing : fog displays a color instead of a texture 	- Two lookup tables to avoid calculus for effects 	- Improved calculus (no div while rendering / int 16.16...) Code normed Added optimization flags in makefile, but not used

+++ srcs/utils/math_utils.c
@@ -6,7 +6,7 @@
 /*   By: arocca <arocca@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/10 15:10:31 by arocca            #+#    #+#             */
-/*   Updated: 2025/10/15 18:47:25 by arocca           ###   ########.fr       */
+/*   Updated: 2025/10/17 11:15:16 by arocca           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ bool	in_bound(int x, int y, int width, int height)
 	return (x >= 0 && x < width && y >= 0 && y < height);
 }
 
-int	clamp(double x, int min, int max)
+double	clamp(double x, int min, int max)
 {
 	if (x < (double)min)
 		return ((double)min);
